requirements.txt
Django~=5.0.1
style~=1.1.0
pip~=24.0
attrs~=23.2.0
#pillow~=10.2.0
asgiref~=3.7.2
sqlparse~=0.4.4
psycopg2~=2.9.9
colorama~=0.4.6
numpy~=1.26.3
requests~=2.31.0
ipython~=8.20.0
typing_extensions~=4.9.0
python-dotenv~=1.0.1
Pygments~=2.17.2
idna~=3.6
certifi~=2024.2.2
telebot~=0.0.5
aiohttp~=3.9.3
urllib3~=2.2.0
chardet~=5.2.0
telegram~=0.0.1
pytz~=2023.4
aiogram~=2.14
django-import-export==3.3.7